requirements.txt
numpy==2.0.2
levenshtein==0.26.1
tqdm==4.67.1
nltk==3.9.1
jieba==0.42.1
sentence-transformers==3.3.1
scikit-learn==1.5.2
faiss-cpu==1.9.0.post1
matplotlib==3.10.0
scipy==1.14.1
torch==2.4.1
transformers==4.47.0
huggingface-hub==0.26.5
openai==1.59.3 